requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-environ==0.4.5
django-modeltranslation==0.15.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2==2.8.5
pyTelegramBotAPI==3.7.2
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
